refactor: reorganize grid test suite
This reorganized the grid test suite into its own test folder under grids. The old file was getting
too long and cumbersome, and there are only going to be more additions soon. This separation of
concerns is more sustainable. Additionally, I added the grids.testdata.js file to export a common
set of data for use amongst the test suite.